bots: Add functionality to test bots by test fixture files.
Initially, if-else block is added so that, adding fixtures
for some files doesn't break others. As test-fixture
file for all bots are added, "test_file_name" will be
asserted to be not None.

Changed parameter name "test_name" to "test_file_name" for
better understanding of the value of the parameter.

Created a common function for reading files, as 2 functions:
mock_http_conversation and check_expected_responses use it.