bots: Modify 'extract_query_without_mention' to take bot name as input.
This is done so that Embedded bot system can also make use of
this function directly, as only client name is needed in this function,
and the type of client need not be ExternalBotHandler.

Also, there would be ambiguity as what type of BotHandler is
passed in the arguments: EmbeddedBotHandler or ExternalBotHandler.