zulip: Change remove_subscriptions to update_subscriptions.
The underlying endpoint that remove_subscriptions called was
zerver.streams.view.update_subscriptions_backend, which is more
for bulk removing *and* adding subscriptions. Also there is a
separate endpoint for removing subscriptions that gives more
control than this one. So this method should be more aptly
called update_subscriptions and should allow one to add
subscriptions as well.

Note that once we make the next PyPI release and update the main
repo to use it, ./zulip/tools/test-api should fail, and would
need to be updated in the PR that updates the main repo to use
the latest release of the packages.