bots/tictactoe: Add simple tests for TicTacToeModel functions.
tictactoe.py did not originally have working tests for the the TicTacToeModel class, these tests were created as a basepoint for future, more complex tests. Originally, we tried to create tests using the built in _get_game_handler() model, but found that this could not easily pass in a hard-coded board. Additionally, we found some issues with the tictactoe.py code itself, as it was not able to recognize the variable current_board because it was not declared as a local variable within the class. In addition to running the automated tests, we ran the tests locally using ./tools/test-bots, and found that a few of the tests that we wrote failed due to an issue with the contains_winning_move() function. Finally, we tested the coverage of our new tests using coverage for Python. We think that special attention should be paid to the change that we made in tictactoe.py, where we set current_move = initial_board (line 30). Additionally, the functions for contains_winning_move() and determine_game_over() should be paid special attention to because these are the tests that the bot itself is failing (most likely an issue with contains_winning_move() function). Finally, for this commit we had to change the code for tools/provision locally on our machine to address the issue with the pip version 10.0.0 that was breaking the dependencies allowing for us to run the virtualenv (referenced in issue #371).

Fixes: #122