bots/tictactoe/test_tictactoe.py: Fixed issues in test cases in test_tictactoe.py.
Encountered several issues with test cases initially, where test cases were failing due to the structure of the tests within the file. This was circumvented by creating a separate instance of the class TicTacToeModel for each test case in order to focus on unit testing. There might be an issue with the _get_game_handlers() function at the end of the file which is used to obtain the model and message information, it seems to be getting the entire class TicTacToeModel instead of creating an instance of the class. Overall, this commit focuses on creating a precedent for writing test cases involving TicTacToeModel, and implements basic cases which can be extended to larger edge cases in the future. Testing was done locally by running ./tools/test-bots until we were able to get the response 'ok' for all the test cases written. We also used coverage in order to test the coverage of the cases. While we were not able to increase the coverage as much as we would have liked, we were able to identify several issues and fix them, hopefully making it easier for future contributors to add additional test cases to TicTacToeModel. Overall, we would like special attention given to the way that TicTacToeModel() was initailized in these tests and whether or not this is a scalable precedent for future tests.

Fixes: #122