zulip_botserver: Handle all requests from the root / endpoint.
Previously, the Botserver determined which bot to run by
dispatching on a different URL endpoint /bots/<botname> for each bot.
Now, instead, the Botserver determines which bot to run by the section
header of the bot in the flaskbotrc.
This implementation simplifies the setup of the Botserver. However,
without the dispatch on the URL endpoint, the Botserver cannot tell
anymore which bot to run for an incoming message. Therefore, it now
inspects each message for the bot name that is @mentioned in it. It
then looks for a section in the flaskbotrc that has this name as a
property and runs the associated bot. The name property is newly
introduced with this commit.