bots: Allow users to supply ZULIP_API_KEY, etc.
This will make it easier to run standalone
bots in containers like Heroku that prefer
env-var-style configuration.

For now this is undocumented, but we should
update the server docs once a few folks have
tried it out.

(The history behind requiring the config file
is that I wanted to keep things simple and
be strongly opinionated about how you run
bots, so that the docs didn't overwhelm folks,
but this use case has come up more frequently.)