integrations: Remove avatar from git commits showing author name instead.
Getting the logs run in cmd as in function `git_commit_range(oldrev, newrev)`
so I had to prefix author name with -[AUTHOR], commit with -[COMMIT]
and subject/title with -[SUBJECT] to be able to extract these info correctly.

I also tried to stick with the format proposed in this issue
https://github.com/zulip/zulip/issues/3968 except for showing the author's
name before each commit and without summary.

Fixes: #632