bot_server: Support trigger private_message renamed to direct_message.
The JSON payload that Zulip server POST for outgoing webhooks
has 'trigger' as one of the fields.

In https://github.com/zulip/zulip/commit/c4e4737, we renamed
the 'private_message' value to 'direct_message'.

This commit adds support to the botserver for handling
'direct_message' as a trigger value. It still supports
'private_message' for self-hosted server compatibility.