google-calendar: Fix usage of term "credentials", replace with "tokens".
Replaces some occurrences of the term "credentials" with "tokens" for
clarity, to conform with the terms used in Google API documentation.

Renamed:
- "google-credentials.json" to "google-tokens.json"
- `credential_path` to `tokens_path`

Google documentation refers to the client secret file sometimes as
"credentials.json", and the tokens file as "tokens.json".
We have been using "client-secret.json" and "google-credentials.json"
respectively, which can be confusing.

So, this commit switches to using the term "tokens" instead of
"credentials" wherever appropriate, to avoid confusion.
The term "credentials" is still used to refer to the Credentials object
returned after authorization.