google-calendar: Call get-google-credentials script internally.
The auth tokens expire every hour. So, the google-calendar script would
stop functioning one hour after start, if it only loads from the token
file.
The get-google-credential script needs to be called directly from the
google-calendar script, without the user needing to run that as a
separate command.

The get-google-credentials script is not a module.
And it uses hyphens in its name, hence it cannot be directly imported
into google-calendar.
To avoid renaming files, and smoothly pass in arguments, runpy is used.

Though google-calendar script is currently the only google service
integration that uses get-google-credentials, the script is not merged
with the google-calendar script, to keep the logic separate, and easy
to expand.

The get-google-credentials script can no longer be run directly, as the
get_credentials() arguments do not have any default values, with the
constants deleted, to avoid redundancy with the google-calendar script.

Updated the function docstrings, usage help and error messages.