google-calendar: Use a bot to send direct messages to the bot owner.
The reminders were previously being sent to the DMs of the current user,
requiring the zuliprc of the user.
But, we can create a generic bot for the Google Calendar integration,
and use its zuliprc to send direct messages to its owner. This tightens
security.

The support for sending DMs to the same user is retained, for cases
where a bot is not used.