google-calendar: Support sending reminders to channels.
Previously, until last commit, the zuliprc of the user was being used
directly. Now that we're using the zuliprc of a bot, it is safe for
multiple users with a shared calendar to have a bot send the reminders
to a channel.

Added channel and topic arguments, with the default behavior set to
sending a direct message.