docs: Do not use relative URL to refer to integrations category.
We are changing individual integration docs to live at
integrations/{integration_name} instead of
integrations/doc/{integration_name}. Our relative link of
`../version-control` in this case will lead to `/version-control` on
the root of the web app instead of intended
`/integrations/version-control`.

We are also changing categories to live at
/integrations/category/{category_name` instead of
/integrations/{category_name} on the web app. But the web app will be
setting up redirects for existing categories to lead them to the new
URL. This commit specifies the /integrations/{category_name} temporarily
so that this PR can be merged before the changes in the web app and we
don't have worry about syncing the releases between the web app and
python-zulip-api.

Once the webapp code is merged, we should change all the relevant links
in this repo to use the new URL structure.