rss-bot: Support unordered RSS feeds.
By splitting the logic into two loops - one for processing all the
entries in the feed, and another to post only the latest ones in
chronological order.

Instead of tracking new_hashes in memory while processing the feed file,
we track unhashed_entries now, since we will not be hashing all the
entries, only the ones that we post.

Fixes #831.